Refactor: change component files location
- containers 폴더 추가
- 동일 컴포넌트에 대한 styled-components 파일과 component 파일을 하나의 폴더에 위치시킴
- components, containers, utils 폴더에 index 파일 추가
- components 내의 파일들 절대경로 적용(Auth 관련 컴포넌트 제외)

close #62 - Refactoring component
