+++ components/MainBanner/main-banner.style.ts
