+++ components/PhotoCardPreview/photocard-preview.style.ts
