+++ components/UserCard/user-card.style.ts
@@ -1,4 +1,4 @@
-import styled from '../styles/themed-components'
+import styled from '../../styles/themed-components'
 
 export const Card = styled.div`
   width: 100%;
