+++ components/index.ts
@@ -1,4 +1,15 @@
-export * from '@components/Buttons'
-export * from '@components/LinkBanner/LinkBanner'
-export * from '@components/LinkTitle/LinkTitle'
-export * from '@components/Searchbars/SearchBar'
+export * from './Buttons'
+export * from './Searchbars/SearchBar'
+export * from './Tag/Tag'
+export * from './Thumbnails/Thumbnail'
+export { default as LinkBanner } from './LinkBanner/LinkBanner'
+export { default as LinkTitle } from './LinkTitle/LinkTitle'
+export { default as UserBanner } from './UserBanner/UserBanner'
+export { default as MainBanner } from './MainBanner/MainBanner'
+export { default as UserCard } from './UserCard/UserCard'
+export { default as PhotoCardPreview } from './PhotoCardPreview/PhotoCardPreview'
+export { default as UserInfoHor } from './UserInfo/UserInfo-hor'
+export { default as UserInfoVer } from './UserInfo/UserInfo-ver'
+export { default as TabMenu } from './TabMenu/TabMenu'
+export { default as LocationInfo } from './LocationInfo/LocationInfo'
+export { default as Layout } from './Layout'
