+++ containers/PhotoCardInput/PhotoCardInput.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
 import { PhotoCardInputContainer, ImageInputbtn } from './photocard-input'
-import LocationInfo from '../LocationInfo/LocationInfo'
+import LocationInfo from '../../components/LocationInfo/LocationInfo'
 import { Button } from '../../styles/common'
 
 interface props {
