+++ containers/index.ts
@@ -1,1 +1,2 @@
-export {}
+export { default as PhotoCardInput } from './PhotoCardInput/PhotoCardInput'
+export { default as PhotoCardGallery } from './PhotoCardGallery/PhotoCardGallery'
