+++ interfaces/user.ts
@@ -1,6 +1,6 @@
 export interface IUser {
   id: number
   name: string
-  imgUrl: string
-  contentCount: number
+  imgUrl?: string
+  contentCount?: number
 }
