--- styles/link-title.ts
@@ -1,9 +1,0 @@
-import styled from './themed-components'
-
-export const LinkTitleContainer = styled.div`
-  font-size: 2rem;
-  font-weight: bold;
-  a::after {
-    content: ' >';
-  }
-`
