+++ tsconfig.json
@@ -21,6 +21,7 @@
     "baseUrl": ".",
     "paths": {
       "@components/*": ["components/*"],
+      "@containers/*": ["containers/*"],
       "@utils/*": ["utils/*"],
       "@hooks/*": ["hooks/*"],
       "@actions/*": ["modules/*"],
