+++ utils/index.ts
@@ -1,0 +1,2 @@
+export * from './sample-data'
+export * from './sample-function'
